Special case of lists now accounted for
`ul > div > li` is not proper html syntax. This change checks to see if
the first item selected is a list. If it is, we eventually fork and
treat the wrapping differently.

Lists are wrapped twice. `ul > li` turns into `ul > li > ul > li`. The
`ul ul` is always an only child, but not classed. Classing, if turned
on, is enabled on the `ul > li li` and s.elem_class and s.elem_id are
applied to the `ul > li` while `s.elem` is ignored.