Wrapping li's no longer clobbers list tag classes
Previously, if a set of li’s was wrapped with ‘ul.classname’ the
‘.classname’ would be lost as the ul tag was removed removed from the
DOM and all li’s were added to new ul tags.

I’m now trying to be intelligent about how this is done. If you have a
list and the ul tag has a class, this overrides the default value
(‘group’) but not anything you enter in. If you enter something in,
it’ll be prepended to the classes which were there. So, $(‘ul >
li’).group(‘.classname’) will result in each of your ul’s having all
their existing classes, plus ‘.classname’ but just calling .group() on
the same selector won’t add ‘.group’